Front end form and firebase database connected
You can now add name, phone, and terminal on the home page form, and
click “Create New User” button to add them to the database. The
database updates immediately with the new user and their fields.

First names and terminals are displayed on the home screen now for
debugging purposes