Photo uploading and displaying locally, photo display from hard-coded database
This code has two parts:
1. User can choose to upload a photo from their phone, and the app will
display it on the page.
2. In the user loop, we now display name, terminal, and imgSrc.
Currently, I hard-coded a base64 string as a field for one of the users
in the firebase browser screen, so one image should display next to the
name of the person who uploaded that image.

TODO:
1. After a user uploads a photo, send the base64 string ($scope.imgSrc)
to the database in the user’s imgSrc field.
2. After a pair of users is matched, display the user’s match’s photo
on the user’s screen so they know who to look for. (Access via
something like user.matchId.imgSrc)